feat: Add requester field to FriendDto
This commit adds a new boolean field called "requester" to the FriendDto data class. The value of this field is set based on whether the user requesting friendship is the same as the logged-in user. This change was made in order to provide more information about friend requests and improve the overall functionality of the application.