feat: Add getFriend endpoint to FriendsResource
This commit adds a new GET endpoint to retrieve a friend of the user. The endpoint returns a FriendDto object and has two possible APIResponses, one for success and another for when the friend is not found. The method calls the getRelationShip method from FriendsService which retrieves the relationship between two users.