feat: Add endpoints to get user friends and relations
This commit adds two new GET endpoints to the FriendsResource class. The getUserFriends endpoint returns a list of friends for a given user ID, while the getRelations endpoint returns a list of relations for a user based on their ID in the header. Additionally, the FriendsRepository class was updated with a new method called getRelations that retrieves all Friend objects where either userRequester or userRequested matches the provided UUID. Finally, the FriendsService class was updated with a new method called getRelations that maps Friend objects returned by FriendsRepository.getRelations to FriendDto objects.